classifier project/requirements.txt
Pillow==9.4.0
tabulate==0.9.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0